pyproject.toml
[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project]
name = "repoindex"
version = "0.10.2"
description = "A collection-aware metadata index for git repositories"
readme = "README.md"
license-files = [ "LICEN[CS]E.*",]
requires-python = ">=3.8"
keywords = [ "git", "github", "repository", "metadata", "index", "cli",]
classifiers = [ "Development Status :: 4 - Beta", "Topic :: Software Development :: Version Control :: Git", "Operating System :: OS Independent", "Programming Language :: Python :: 3",]
dependencies = [ "pathlib", "toml", "requests>=2.25.0", "packaging>=21.0", "click>=8.0.0", "rich>=13.0.0", "tweepy", "rapidfuzz", "pyyaml>=5.0.0",]
[[project.authors]]
name = "Alex Towell"
email = "lex@metafunctor.com"

[project.license]
text = "MIT"

[project.optional-dependencies]
test = [ "pytest>=6.0", "pytest-cov>=2.0", "coverage>=5.0",]
dev = [ "pytest>=6.0", "pytest-cov>=2.0", "coverage>=5.0", "black>=21.0", "flake8>=3.9", "mypy>=0.812",]
docs = [ "mkdocs>=1.4.0", "mkdocs-material>=9.0.0", "ghp-import>=2.1.0",]
export = [ "weasyprint>=54.0",]
network = [ "numpy>=1.20.0",]
clustering = [ "numpy>=1.20.0", "scikit-learn>=1.0.0", "scipy>=1.7.0",]
workflow = [ "pyyaml>=5.0.0",]
all = [ "numpy>=1.20.0", "scikit-learn>=1.0.0", "scipy>=1.7.0", "weasyprint>=54.0",]

[project.scripts]
repoindex = "repoindex.cli:main"

[project.urls]
Homepage = "https://queelius.github.io/repoindex"
Documentation = "https://queelius.github.io/repoindex"
Source = "https://github.com/queelius/repoindex"
Issues = "https://github.com/queelius/repoindex/issues"

[tool.setuptools]
packages = [ "repoindex", "repoindex.commands", "repoindex.integrations", "repoindex.providers", "repoindex.exporters",]

[tool.setuptools.package-data]
repoindex = [ "py.typed",]

[tool.setuptools.exclude-package-data]
"*" = [ "tests", "tests.*", "*.tests", "*.tests.*", "test_*",]
